Split template.rs into smaller submodules
