+++ src/template/parse.rs
@@ -1,275 +1,23 @@
-use std::fmt;
-use std::ops::{Deref, Index};
-
 use color_eyre::eyre::{eyre, Result};
 
-use crate::variables::{UserVars, Variables};
+use super::span::{CharSpan, Spanned};
+use super::token::{If, IfExpr, IfOp, Token, Var, VarEnv, VarEnvSet};
+use super::Template;
 
-// TODO: handle unicode
-
-type CharPosType = u32;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct CharPos(CharPosType);
-
-impl CharPos {
-	pub fn from_usize(value: usize) -> Self {
-		Self(value as CharPosType)
-	}
-
-	pub fn as_usize(&self) -> usize {
-		self.0 as usize
-	}
-}
-
-impl From<usize> for CharPos {
-	fn from(value: usize) -> Self {
-		CharPos::from_usize(value)
-	}
-}
-
-impl From<CharPosType> for CharPos {
-	fn from(value: CharPosType) -> Self {
-		Self(value)
-	}
-}
-
-impl From<CharPos> for usize {
-	fn from(value: CharPos) -> Self {
-		value.as_usize()
-	}
-}
-
-impl From<CharPos> for CharPosType {
-	fn from(value: CharPos) -> Self {
-		value.0
-	}
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct CharSpan {
-	low: CharPos,
-	high: CharPos,
-}
-
-impl CharSpan {
-	pub fn new<L: Into<CharPos>, H: Into<CharPos>>(low: L, high: H) -> Self {
-		let mut low = low.into();
-		let mut high = high.into();
-
-		if low > high {
-			std::mem::swap(&mut low, &mut high);
-		}
-
-		Self { low, high }
-	}
-
-	pub fn span<T>(self, value: T) -> Spanned<T> {
-		Spanned::new(self, value)
-	}
-}
-
-impl Index<CharSpan> for str {
-	type Output = str;
-
-	fn index(&self, index: CharSpan) -> &Self::Output {
-		&self[index.low.as_usize()..index.high.as_usize()]
-	}
-}
-
-impl Index<&CharSpan> for str {
-	type Output = str;
-
-	fn index(&self, index: &CharSpan) -> &Self::Output {
-		&self[index.low.as_usize()..index.high.as_usize()]
-	}
-}
-
-pub struct Spanned<T> {
-	span: CharSpan,
-	value: T,
-}
-
-impl<T> Spanned<T> {
-	fn new(span: CharSpan, value: T) -> Self {
-		Self { span, value }
-	}
-
-	fn span(&self) -> &CharSpan {
-		&self.span
-	}
-
-	fn value(&self) -> &T {
-		&self.value
-	}
-
-	fn into_span(self) -> CharSpan {
-		self.span
-	}
-
-	fn into_value(self) -> T {
-		self.value
-	}
-
-	fn into_inner(self) -> (CharSpan, T) {
-		(self.span, self.value)
-	}
-}
-
-impl<T> fmt::Debug for Spanned<T>
-where
-	T: fmt::Debug,
-{
-	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-		f.debug_struct("Spanned")
-			.field("span", &self.span)
-			.field("value", &self.value)
-			.finish()
-	}
-}
-
-impl<T> Clone for Spanned<T>
-where
-	T: Clone,
-{
-	fn clone(&self) -> Self {
-		Self {
-			span: self.span,
-			value: self.value.clone(),
-		}
-	}
-}
-
-impl<T> Copy for Spanned<T> where T: Copy {}
-
-impl<T> PartialEq for Spanned<T>
-where
-	T: PartialEq,
-{
-	fn eq(&self, other: &Self) -> bool {
-		self.span.eq(&other.span) && self.value.eq(&other.value)
-	}
-}
-
-impl<T> Eq for Spanned<T> where T: Eq {}
-
-impl<T> Deref for Spanned<T> {
-	type Target = T;
-
-	fn deref(&self) -> &Self::Target {
-		&self.value
-	}
-}
-
-#[derive(Debug, Clone, PartialEq)]
-enum Token {
-	Var(Var),
-	If(If),
-	Escaped(CharSpan),
-	Comment,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Hash)]
-enum VarEnv {
-	Environment,
-	Profile,
-	Item,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Hash)]
-struct VarEnvSet([Option<VarEnv>; 3]);
-
-impl VarEnvSet {
-	fn empty() -> Self {
-		Self([None; 3])
-	}
-
-	fn add(&mut self, value: VarEnv) -> bool {
-		if self.0.contains(&Some(value)) {
-			false
-		} else if let Some(slot) = self.0.iter_mut().find(|x| x.is_none()) {
-			*slot = Some(value);
-			true
-		} else {
-			false
-		}
-	}
-
-	fn envs(&self) -> impl Iterator<Item = &VarEnv> {
-		self.0.iter().filter_map(|x| x.as_ref())
-	}
-
-	fn len(&self) -> usize {
-		self.envs().count()
-	}
-
-	fn capacity(&self) -> usize {
-		self.0.len()
-	}
-}
-
-impl Default for VarEnvSet {
-	fn default() -> Self {
-		Self([
-			Some(VarEnv::Item),
-			Some(VarEnv::Profile),
-			Some(VarEnv::Environment),
-		])
-	}
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Hash)]
-struct Var {
-	envs: VarEnvSet,
-	name: CharSpan,
-}
-
-// For now only eq.
-#[derive(Debug, Clone, PartialEq)]
-struct If {
-	head: Spanned<IfExpr>,
-	elifs: Vec<Spanned<IfExpr>>,
-	els: Option<CharSpan>,
-	end: CharSpan,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Hash)]
-enum IfOp {
-	Eq,
-	NotEq,
-}
-
-impl IfOp {
-	fn eval(&self, lhs: &str, rhs: &str) -> bool {
-		match self {
-			Self::Eq => lhs == rhs,
-			Self::NotEq => lhs != rhs,
-		}
-	}
-}
-
-// For now only eq.
-#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Hash)]
-struct IfExpr {
-	var: Var,
-	op: IfOp,
-	other: CharSpan,
-}
-
-struct Parser<'a> {
+pub struct Parser<'a> {
 	content: &'a str,
 	blocks: BlockIter<'a>,
 }
 
 impl<'a> Parser<'a> {
-	fn new(s: &'a str) -> Self {
+	pub fn new(s: &'a str) -> Self {
 		Self {
 			content: s,
 			blocks: BlockIter::new(s),
 		}
 	}
 
-	fn parse(mut self) -> Result<Template<'a>> {
+	pub fn parse(mut self) -> Result<Template<'a>> {
 		let tokens =
 			std::iter::from_fn(|| self.parse_next_block()).collect::<Result<Vec<_>, _>>()?;
 
@@ -286,19 +34,19 @@ impl<'a> Parser<'a> {
 		println!("Content: `{}`", content);
 
 		if let Some(b'{') = content.as_bytes().get(0) {
-			let escaped = CharSpan::new(span.low.as_usize() + 3, span.high.as_usize() - 3);
-			return Some(Ok(Spanned::new(span, Token::Escaped(escaped))));
+			let escaped = span.offset_low(3).offset_high(-3);
+			return Some(Ok(span.span(Token::Escaped(escaped))));
 		}
 
 		if let Some(b"!--") = content.as_bytes().get(..3) {
-			return Some(Ok(Spanned::new(span, Token::Comment)));
+			return Some(Ok(span.span(Token::Comment)));
 		}
 
 		if let Some(b"@if") = content.as_bytes().get(..3) {
 			return Some(self.parse_if_blocks(span.span(content)));
 		}
 
-		match parse_var(content, span.low.as_usize() + 2) {
+		match parse_var(content, span.low().as_usize() + 2) {
 			Ok(var) => Some(Ok(span.span(Token::Var(var)))),
 			Err(err) => Some(Err(err)),
 		}
@@ -313,18 +61,17 @@ impl<'a> Parser<'a> {
 		let head = {
 			let start = content
 				.find("{{")
-				.ok_or_else(|| eyre!("Found no variable block in if at {:?}", span.low))?;
+				.ok_or_else(|| eyre!("Found no variable block in if at {}", span))?;
 			let end = content
 				.find("}}")
-				.ok_or_else(|| eyre!("Found no closing for variable block in if at {:?}", start))?;
+				.ok_or_else(|| eyre!("Found no closing for variable block in if at {}", span))?;
 
 			// +2 for if opening `{{` the other +2 for var opening
-			let var = parse_var(
-				&content[start + 2..end],
-				span.low.as_usize() + 2 + start + 2,
-			)?;
+			let var_start = span.low().as_usize() + 2 + start + 2;
+
+			let var = parse_var(&content[start + 2..end], var_start)?;
 			let op = parse_ifop(&content[end + 2..])?;
-			let other = parse_other(&content[end + 2..], span.low.as_usize() + 2 + end + 2)?;
+			let other = parse_other(&content[end + 2..], var_start)?;
 
 			span.span(IfExpr { var, op, other })
 		};
@@ -347,21 +94,19 @@ impl<'a> Parser<'a> {
 
 			if let Some(b"@elif") = content.as_bytes().get(..5) {
 				let elif = {
-					let start = content.find("{{").ok_or_else(|| {
-						eyre!("Found no variable block in elif at {:?}", span.low)
-					})?;
+					let start = content
+						.find("{{")
+						.ok_or_else(|| eyre!("Found no variable block in elif at {}", span))?;
 					let end = content.find("}}").ok_or_else(|| {
-						eyre!("Found no closing for variable block in elif at {:?}", start)
+						eyre!("Found no closing for variable block in elif at {}", span)
 					})?;
 
 					// +2 for if opening `{{` the other +2 for var opening
-					let var = parse_var(
-						&content[start + 2..end],
-						span.low.as_usize() + 2 + start + 2,
-					)?;
+					let var_start = span.low().as_usize() + 2 + start + 2;
+
+					let var = parse_var(&content[start + 2..end], var_start)?;
 					let op = parse_ifop(&content[end + 2..])?;
-					let other =
-						parse_other(&content[end + 2..], span.low.as_usize() + 2 + end + 2)?;
+					let other = parse_other(&content[end + 2..], var_start)?;
 
 					span.span(IfExpr { var, op, other })
 				};
@@ -409,7 +154,7 @@ impl<'a> Parser<'a> {
 
 		let end = span;
 
-		let whole_span = CharSpan::new(head.span().low, end.high);
+		let whole_span = head.span().union(&end);
 
 		Ok(whole_span.span(Token::If(If {
 			head,
@@ -417,140 +162,6 @@ impl<'a> Parser<'a> {
 			els,
 			end,
 		})))
-	}
-}
-
-#[derive(Debug, Clone)]
-pub struct Template<'a> {
-	content: &'a str,
-	tokens: Vec<Spanned<Token>>,
-}
-
-impl<'a> Template<'a> {
-	pub fn parse(content: &'a str) -> Result<Self> {
-		Parser::new(content).parse()
-	}
-
-	// TODO: trim `\r\n` when span start/ends with it
-	pub fn fill(
-		&self,
-		profile_vars: Option<&UserVars>,
-		item_vars: Option<&UserVars>,
-	) -> Result<String> {
-		let mut output = String::new();
-		let mut idx = 0;
-
-		for Spanned { span, value: token } in &self.tokens {
-			if idx < span.low.as_usize() {
-				output.push_str(&self.content[idx..span.low.as_usize()])
-			}
-
-			match token {
-				Token::Var(var) => {
-					output.push_str(&self.resolve_var(var, profile_vars, item_vars)?);
-				}
-				Token::If(If {
-					head,
-					elifs,
-					els,
-					end,
-				}) => {
-					let head_val = self.resolve_var(&head.var, profile_vars, item_vars)?;
-					if head.op.eval(&head_val, &self.content[head.other]) {
-						let span = CharSpan::new(
-							head.span().high.as_usize(),
-							elifs
-								.first()
-								.map(|elif| elif.span())
-								.unwrap_or_else(|| els.as_ref().unwrap_or(end))
-								.low
-								.as_usize(),
-						);
-						output.push_str(&self.content[span]);
-					} else {
-						let mut found = false;
-						for idx in 0..elifs.len() {
-							let Spanned {
-								span,
-								value: IfExpr { var, op, other },
-							} = &elifs[idx];
-
-							let elif_val = self.resolve_var(var, profile_vars, item_vars)?;
-
-							if op.eval(&elif_val, &self.content[other]) {
-								let span = CharSpan::new(
-									span.high.as_usize(),
-									elifs
-										.get(idx + 1)
-										.map(|elif| elif.span())
-										.unwrap_or_else(|| els.as_ref().unwrap_or(end))
-										.low
-										.as_usize(),
-								);
-								output.push_str(&self.content[span]);
-								found = true;
-							}
-						}
-
-						if !found {
-							if let Some(span) = els {
-								let span = CharSpan::new(span.high.as_usize(), end.low.as_usize());
-								output.push_str(&self.content[span]);
-							}
-						}
-					}
-				}
-				Token::Escaped(inner) => {
-					output.push_str(&self.content[inner]);
-				}
-				Token::Comment => {
-					// NOP
-				}
-			};
-
-			idx = span.high.as_usize();
-		}
-
-		if idx < self.content.len() {
-			output.push_str(&self.content[idx..]);
-		}
-
-		Ok(output)
-	}
-
-	fn resolve_var(
-		&self,
-		var: &Var,
-		profile_vars: Option<&UserVars>,
-		item_vars: Option<&UserVars>,
-	) -> Result<String> {
-		let name = &self.content[var.name];
-
-		for env in var.envs.envs() {
-			match env {
-				VarEnv::Environment => {
-					if let Ok(val) = std::env::var(name) {
-						return Ok(val);
-					}
-				}
-				VarEnv::Profile => {
-					if let Some(Some(val)) = profile_vars.map(|vars| vars.var(name)) {
-						return Ok(val.to_string());
-					}
-				}
-				VarEnv::Item => {
-					if let Some(Some(val)) = item_vars.map(|vars| vars.var(name)) {
-						return Ok(val.to_string());
-					}
-				}
-			};
-		}
-
-		Err(eyre!(
-			"Failed to resolve variable `{}` (Envs: {:?})",
-			name,
-			var.envs
-		))
 	}
 }
 
@@ -724,11 +335,8 @@ impl<'a> Iterator for BlockIter<'a> {
 	fn next(&mut self) -> Option<Self::Item> {
 		let mut span = find_block(&self.content[self.index..])?.unwrap();
 
-		// TODO: make better
-		span.low.0 += self.index as CharPosType;
-		span.high.0 += self.index as CharPosType;
-
-		self.index = span.high.as_usize();
+		span = span.offset(self.index as i32);
+		self.index = span.high().as_usize();
 
 		Some((span, &self.content[span]))
 	}
@@ -736,9 +344,8 @@ impl<'a> Iterator for BlockIter<'a> {
 
 #[cfg(test)]
 mod tests {
-	use std::collections::HashMap;
-
 	use super::*;
+	use crate::template::span::{CharPos, CharSpan};
 
 	#[test]
 	fn find_blocks() {
@@ -868,55 +475,6 @@ mod tests {
 
 		assert!(parse_other("This is a test \"Hello World How are you today", 0).is_err());
 		assert!(parse_other("This is a test", 0).is_err());
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_template() -> Result<()> {
-		let content = r#"
-			[some settings]
-			var = 2
-			foo = "bar"
-			fizz = {{BUZZ}}
-			escaped = {{{42}}}
-
-			{{!--
-				Sets the message of the day for a specific operating system
-				If no os matches it defaults to a generic one.
-			--}}
-			{{@if {{&OS}} == "linux" }}
-			[linux]
-			motd = "very nice"
-			{{@elif {{&#OS}} == "windows" }}
-			[windows]
-			motd = "nice"
-			{{@else}}
-			[other]
-			motd = "who knows"
-			{{@fi}}
-
-			{{!-- Check if not windows --}}
-			{{@if {{&OS}} != "windows"}}
-			windows = false
-			{{@fi}}
-
-			[last]
-			num = 23
-			threads = 1337
-			os_str = "_unkown"
-			"#;
-
-		let template = Template::parse(content)?;
-
-		println!("{:#?}", template);
-
-		let mut vars = HashMap::new();
-		vars.insert(String::from("BUZZ"), String::from("Hello World"));
-		vars.insert(String::from("OS"), String::from("linux"));
-		let vars = UserVars { inner: vars };
-
-		println!("{}", template.fill(Some(&vars), Some(&vars))?);
 
 		Ok(())
 	}
